Move downloads page to version list dropdown menu
This adds a downloads list to the version list with the built formats.
This interaction is API driven as all of the URL lookups in our resolver
are rather expensive.

This adds a view class that expects a URL to hit for api data, and child
classes manipulate the resulting data via knockout subscription or
bindings.