Use `get_static_prefix` instead of `webpack_static`
`webpack_static` fails because it finds a specific file in the manifest and it
does not found it because we are passing a path.

`get_static_prefix` does what we want here: return the prefix for the static
path where the static assets are served from. Locally, with this change the path
returned is `http://community.dev.readthedocs.io/static/readthedocsext/theme`
which is the same returned by `webpack_static`