Remove some obfuscated magic for SUI modules
Things like popup, modal, dropdown, etc have jquery plugins to enable
features on the elements. There are a number of cases where it is
important to have special configuration, so we previously had some
strange selector logic to try to get these edge cases.

Now, everything is explicit in the template code, except for very basic
usage. The Knockout binding `semanticui` is what enables these jQuery
plugins, and it sets an attribute on the element so that the automatic
defaults aren't applied to the elements.