Finish up on the docs for JS and templates
This uses sphinx-js and jsdoc to extract docs from the JS sources, and
autoanysrc to extract docs from the templates. This tries to explain
some of the interaction and patterns used between JS and the templates
as well, as there are a few pieces of glue to make the application and
front end code work well together.