Swap SUI provided Font Awesome for FA6 Pro via hosted kit (#108)
* Replace SUI Font Awesome with FA pro kit

This reduces the duplicate definitions for icons and replaces the
loading mechanism with FA kits, which are a good option to keep load
times lower.

* Swap SUI provided Font Awesome for FA6 Pro via hosted kit

The hosted kit from Font Awesome provides optimized and auto-subsetted
version of the webfont downloads, greatly reducing the footprint of the
Font Awesome.

This follows site-community and drops the included FA5 using a few
hacks. It also drops all the SUI `.icon` class definitions, saving more
bandwidth. All of the icons in templates were replaced with a mix of
solid and duotone icons.

* Update built artifacts