Update project header bar
This will definitely need a second pass, but at least includes most of
the elements that we'll want in the header. The navigation menu is more
strictly for elements directly related to Project, and moves mocked
linkes for translations and subprojects to header elements that aren't
always showing on other Project views.