Example middleware to gate beta access
This doesn't require any view changes. Not a super clean implementation,
as this will allow users to log in (the auth backend didn't change), but
will just continually redirect to the login page. This seems fine
though.

Implementation on the application would be adding to middlewares:

```
if self.RTD_EXT_THEME_ENABLED:
    middlewares.append('readthedocsext.theme.middleware.BetaAuthenticationMiddleware')
```