Update icon import pattern to work with FUI 2.9.2
FUI 2.9.2 changed the import pattern for icon fonts, and started
building these again. We don't use them however, and rely on FA kits
instead. This uses a fix on the common theme to avoid importing these
fonts again, also moving the dependency to FUI to the common theme
instead of here.