Add project redirects templates back
These moved off to the redirects app for some reason, moved them back to
the projects app for the project dashboard and fixed a number of display
issues with the listing and form templates.

I updated this to use the table list display and made the information
structure a bit nicer. There is a big bug with the edit form still,
which looks to be the KO code interfering. I think this code was written
before this was editable :grimacing:

Fixes #133