Wrap table listing headers nicely
Instead of trying to hide long content lines in the table list view,
this instead tries to wrap header content in a nicer way. Trying to clip
the header content was difficult due to how tables work.

There are some base template changes to allow for more control of the
columns and classes used in the table listing, but the header classes
ended up being the only one I needed.

Closes #156