Fix log in and sign up page UX issues (#174)
* Add requirements for using tab module correctly

Adds dependency for tab history, jquery-address, and a helper module for
consolidating the tab module instantiation call from our semanticui data
binding.

* Lint

* Rebuild assets

* Rework log in and sign up forms slightly

This extends the sign up form from the log in form to ensure the
structure is the same between the two pages. A few fixes are in the log
in form page, namely some work around the tab module usage so that we
can link directly to the email log in and email sign up tabs when
needed. This is required to load the correct tab page on form errors,
and we instruct forms to hit the form URL with the tab history hash in
the URL. This is the piece that makes the module detect the active tab
page.

* Rebuild assets

* Rebuild