Add dynamic avatar view for project avatars
I waffled on this one a bit, but did an initial pass inside this
application instead of the core application. I'm impartial here though.

This adds a URL and view to a dynamic view that:

- Redirects to the project's remote repository avatar URL if it exists
- Shows a generated avatar SVG from a template otherwise

This replaces the use of Dicebear, a third party API and site. Dicebear
was added as a placeholder for some nicer images.

It also uses a different style for the avatars, using a reimplementation
of the Dicebear "initials" API instead of the "shapes" API. The shapes
look pretty but they don't feel identifiable either. The project avatar
breaks up the project name/slug into words and makes and acronym from
that.