Add unified build list and build detail page title for builds
This addresses a few issues from #158. The build name used on the build
listing and build detail page both differed prior, and didn't mention
the version name consistently.

- Build detail page just used the build pk
- Build listing page listed all versions as "Version {{ version_name }}",
  even pull request builds, which were "Version 1234" in the build
  listing

Additionally, this adds some hierarchy to the build detail page, showing
an additional layer of nesting for both the version (links to a filtered
version listing) and the build list (links to a filtered build listing).

Closes #158