Use remoterepo v3 API response in project creation view
With the v3 remote repo API additions, this can be swapped to the new
API. This gives more project attributes, which we'll eventually use more
here, but even just the project name is a big improvement.

I opened #196 to consider an implementation of the popup cards that
would be purely API/JS driven, but that is a small improvement to this
UI. The project matches probably shouldn't show up too often, unless you
are mananging a lot of project translations or something redundant like
that.