Add test for Knockout semanticui data binding
This confirms that the custom binding for SUI modules works as expected
by inspecting the element and confirming it seems to have had our logic
and the jQuery SUI plugin logic applied to it.

This found a small bug with the binding code itself. I only tested
dropdown for now, I'm not sure if it's worth testing anything too deeply
yet.