Drop test for now
I couldn't get the modal to detect as active using the internal data
structure added by the FUI module JS. I suspect this requires some CSS,
or perhaps just some missing JS. This could be an issue with just our
test suite, but I won't spend time with this unless we are reusing this
pattern more. I believe we should remove this pattern entirely, and
replace it with proper CreateView and dedicated endpoints.