Add template and model view changes for organization authorization settings
This relies on https://github.com/readthedocs/readthedocs-corporate/pull/1730

This drops the very custom form for a more native form, but retains some
of the styles of the authorization provider list. This is now just a
informational block describing the providers.

This also adds the KO view logic back to the view, so that we can
conditionally show/require the domain field.

Also raised in the PR above, a warning block is added when the provider
value is changing, so there is some information that
authentication/authorization might fail.