Add project: highlight yaml examples
I used `highlight.js` to highlight the examples when importing a project.
It requires importing a CSS file that I'm not sure where to put it yet.

https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.9.0/styles/dark.min.css

Besides, I need to know where to place the small chunk of JS code that I added
to make this work.

Based on #368