Organization: create workflow
Initial approach to improve the "Organizaion creation workflow" and help users
to remove confusions when signing up.

The idea is to expose the 3 alternatives clearly and let them decide what's the
step they want to follow:

- join an existing organization
- create a new organization
- not sure what I'm doing

Depending what option they select, we tell them to contact the organization
owner, we present a form to create an organization or we tell them to contact us.

Related https://github.com/readthedocs/readthedocs.org/issues/11356