Add project: show more `yaml` examples (#368)
* Add project: show more `yaml` examples

I used tabs to show different examples of YAML files, as suggested in #184.
It looks great for an initial interactive approach.

I like the tabs over the dropdown because it gives the user a clear idea there
are more than Sphinx and MkDocs tools supported, which is a good marketing
strategy, in my opinion.

It also adds an "Others" option that's more generic and shows how to use
`build.commands` in a generic way: install dependencies, build, copy assets
under `$READTHDOCS_OUTPUT` directory.

* Update readthedocsext/theme/templates/projects/import_config.html

Co-authored-by: Anthony <aj@ohess.org>

* Update readthedocsext/theme/templates/projects/import_config.html

Co-authored-by: Anthony <aj@ohess.org>

* Combine Dropdown + Tabs

* Fix indentation

* More indentation

---------

Co-authored-by: Anthony <aj@ohess.org>