Update patterns, make a standalone WC for the menu button
This is because the redirect URL depends on the data response structure
and using JS works best here. It does mean that if we add a API wrapper
for something else -- deleting a version through the API or whatever --
we will need a web component specific for that API interaction.