Consolidate patterns for build retry UI elements (#381)
* Consolidate patterns for build retry UI elements

There were a number of ways our dashboard was allowing rebuilding a
version/build. Some used hidden forms, and the external build retry used
different logic from standard builds.

Instead, this makes a web component all with the same API call. This
needs a patch to the API endpoint to trigger both standard and external
version builds.

- Requires https://github.com/readthedocs/readthedocs.org/pull/11407
- Fixes #336

* Update patterns, make a standalone WC for the menu button

This is because the redirect URL depends on the data response structure
and using JS works best here. It does mean that if we add a API wrapper
for something else -- deleting a version through the API or whatever --
we will need a web component specific for that API interaction.