Add templates for Allauth MFA
I placed this as a subpage under the profile "Security" admin dashboard
for now, though this is still not linked anywhere.

The entrance pages use the minimal login base templates, while the
management pages use the user profile edit base as a base template, so
that all of the management functions appear under the user profile
admin.

- Fixes #443