Use multiple tabs for addons configuration page
This is the same as using Crispy Layout API, but allows for authoring
more content in the tabs as we need. For example, we might want to
include placehold segements or images in the tabs.

One improvement might be a way to define the fields for each tab in the
Form instance, so that we don't have to hardcode field names in templates,
but instead iterate over each field in `form.fieldsets.flyout`, or
similar.