Replace project and version KO views (#532)
This solves a few issues and explores a pattern that we need to figure
out in
order to replace Knockout views with web components: authoring and
connection
between nested web components. This is a low stakes place to try this.

The issues around the buttons were addressed through adding a click
wrapper to
the link. If you accidentally trigger the link before the API response,
the
click event is stored and re-emitted after the response finishes.

The pattern being explored here is using the web component context API.
This is
implemented in Lit already. You can read more here:

- https://lit.dev/docs/data/context/
-
https://github.com/webcomponents-cg/community-protocols/blob/main/proposals/context.md

So far, we have been trying out manually using events to pass around
data at
a global level. The context API functions similar under the hood, but
the API is
a much cleaner way to pass data around, and it functions only between
parent and
child elements, not globally.

Where this differs from global events is that we can use this pattern
between
parent and child elements, and we don't need to manually search the DOM
for the
parent element:

```html
<readthedocs-api url="...">
  <div>
    <div>
      <div>
        <readthedocs-item-docs>
          This element will traverse up the tree and listen to only the first
          readthedocs-api provider
        </readthedocs-item-docs>
      </div>
    </div>
  </div>
</readthedocs-api url="...">
<readthedocs-api url="...">
  <readthedocs-item-docs>
    And this element will only listen for data provided by that second
    readthedocs-api provider
  </readthedocs-item-docs>
</readthedocs-api>
```

## Why?

We want to author as little as we need to in JS, and as much as we can
in
templates. Using the context API allows us to author isolated web
components
that share data _while still using Django templates for as much of our
page
structure as we can_.

The alternative is authoring everything nested in `readthedocs-api` in a
big
web component, which is not a good pattern for a Django backend.

The work here so far this is a very basic example and the
provider/consumer
elements are close in proximity in the DOM. A more complex example would
be a
table listing, where the table row is a data provider and elements
nested
anywhere in that row element can share data.

We want to keep as much of that row in Django templates as possible,
especially
if that row is 95% vanilla HTML elements and 5% dynamic web components.

----

Things I did not do here:

- Deprecate the older elements using the same base
- Remove old project/version KO views
- Split up code

This will come next.

----

- Fixes #444
- Fixes #497