Add initial Settings Overview page
This gives users a bit more context about their project,
and is a starting point to introduce more information.

This is a first draft, but curious to get some feedback on what should be included.
I'm a little worried about the performance in production, so will put this behind a staff flag for now.