Clean up migration project listing (#579)
- Use partials instead of one off listings
- Reduce text and move to UI elements
- Compact information in table listing and move warnings about
installation/migration issues to the button. Show button as disabled
when next steps are not possible.
- Use a partial for the second list at the bottom

Relevant commit is be49f17


![image](https://github.com/user-attachments/assets/07c6ee77-0390-4866-8d13-ae474ec13bd2)

This addresses issues with the UI being quite cramped and elements
overlapping/wrapping and the UI generally having a bit too much text.

I did not test if pagination would work, I assume not. I don't love
having two lists on this UI, it would be better as a singular list
instead of two. This would require reworking the logic used, namely the
use of dataclasses for each list of objects.