Add migration page for GitHub App (#570)
While this still requires
https://github.com/readthedocs/readthedocs.org/pull/12112, I think it
should be safe to review, as the UI shouldn't change much. ~~I still
need to finishe a couple of things: make all strings translable, and
check for users with multiple GH accounts connected~~.

- Step by step guide to connect your account to the new GH app, and also
to migrate projects.
- Partial support for users with more than one GH account linked (they
just need to repeat the migration twice), we don't have a lot of users
with more than one account connected, so this should be fine.
- Automatically refresh the page once the GH tab has been closed, this
should work for most cases, but it won't work if the user doesn't close
the page, or if the page refreshes before our application was able to
sync the repos. We still show a message about refreshing the page, so.


https://github.com/user-attachments/assets/a836dfaf-4f30-41ca-a84b-8a669c838aef

---------

Co-authored-by: Anthony <aj@ohess.org>