Add Visit docs button to header (#508)
This gives a basic way to get to the docs like the old dashboard. I was
doing a bunch of local dev on the new dashboard and kept getting annoyed
at not being able to get to the docs in an obvious way.

I think this is important and is a middle ground until we can work on a
fancier verison.

<img width="939" alt="Screenshot 2024-10-15 at 7 21 13 AM"
src="https://github.com/user-attachments/assets/a701d20e-3965-4232-9360-7a93795f41ec">