GitHub App: open beta (#603)
The revoke and disconnect options are not available during the beta, so
users can use the old button to log in without any problems. They can
still do this from the social accounts list, but if they do so, next
time they log in will be redirected to the correct login page, so.

Ref https://github.com/readthedocs/meta/issues/187