Add project create modal for repairing the repository listing (#648)
This is a stronger UI that adds a little more content around why the
user might need to resync/reconfigure permissions. This isn't yet as
aggressive as it might need to be, it's just the scaffolding for the UI
and experimenting a bit more with the UX and feedback to the user.

- Fixes #620 
- Requires readthedocs/readthedocs.org#12506