Block GHA migration view for SSO organization users
This shows the user a verbose error message and blocks the view from
being used for now, and points the user to contact support. From there,
we can either help the user manually migrate or we can at least explain
the limitation and follow up when we are ready to test support for
dual-integration permission checks.