Use better patterns for login view (#671)
My feeback on what the LLM did to #622 is that it should use Knockout
and SUI
instead of trying to continue pushing the usage into template and manual
element
manipulation with JS.

This solves the UX issues with popups, SUI popups are stronger when used
directly instead of using them HTML only.

I didn't test this thoroughly.



https://github.com/user-attachments/assets/e8dd620c-ac17-4cab-9281-39793ecfc403